pyproject.toml
[project]
name = "kitchen-hygiene-monitor-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.6",
    "opencv-python>=4.13.0.92",
    "python-telegram-bot>=22.5",
]
